refactor: extract _subprocess.py from helpers.py and migrate consumers
Move _run_subprocess, _process_runner_result, _GH_API_SUBCOMMANDS, and
_is_github_cli_call into server/_subprocess.py. Retarget all 6 tool files
and 2 test files to import from the new location. Remove extracted symbols
and orphaned imports from helpers.py.
